feat(table/date-cell-popover): refactor code
